Allow disabling of loop callback for modules

+++ ESPGeiger/src/Module/EGModuleRegistry.h
@@ -41,7 +41,8 @@ class EGModuleRegistry {
 #endif
     static void log_activity_and_reset();
     static void wake();
-    static bool set_loop_interval(EGModule* m, uint16_t interval_ms);
+    // <0 disables loop, 0 = every iteration, positive = ms (clamped to 65535).
+    static bool set_loop_interval(EGModule* m, int32_t interval_ms);
     static bool sleep_until(EGModule* m, unsigned long now, unsigned long target_ms);
     static uint8_t count();
     static EGModule* get(uint8_t idx);
